+++ app/src/main/java/com/konkuk/medicarecall/domain/model/ElderHealthInfo.kt
@@ -4,7 +4,7 @@ import com.konkuk.medicarecall.domain.model.type.HealthIssueType
 import com.konkuk.medicarecall.domain.model.type.MedicationTime
 
 data class ElderHealthInfo(
-    val elderId: Int,
+    val elderId: Long,
     val name: String,
     val diseases: List<String> = emptyList(),
     val medications: Map<MedicationTime, List<String>> = emptyMap(),
